refactor: application lists get data from prop
