+++ frontend/benefit/handler/src/constants.ts
@@ -182,3 +182,12 @@ export const APPLICATION_INITIAL_VALUES = {
   applicationStep: DEFAULT_APPLICATION_STEP,
   [APPLICATION_FIELD_KEYS.APPLICATION_ORIGIN]: APPLICATION_ORIGINS.HANDLER,
 };
+
+export const ALL_APPLICATION_STATUSES: APPLICATION_STATUSES[] = [
+  APPLICATION_STATUSES.RECEIVED,
+  APPLICATION_STATUSES.HANDLING,
+  APPLICATION_STATUSES.INFO_REQUIRED,
+  APPLICATION_STATUSES.ACCEPTED,
+  APPLICATION_STATUSES.DRAFT,
+  APPLICATION_STATUSES.REJECTED,
+];
