+++ frontend/benefit/handler/src/pages/index.tsx
@@ -4,6 +4,7 @@ import MainIngress from 'benefit/handler/components/mainIngress/MainIngress';
 import AppContext from 'benefit/handler/context/AppContext';
 import FrontPageProvider from 'benefit/handler/context/FrontPageProvider';
 import { APPLICATION_STATUSES } from 'benefit-shared/constants';
+import { ApplicationListItemData } from 'benefit-shared/types/application';
 import { Tabs } from 'hds-react';
 import { GetStaticProps, NextPage } from 'next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
@@ -13,16 +14,9 @@ import Container from 'shared/components/container/Container';
 import theme from 'shared/styles/theme';
 
 import { useApplicationList } from '../components/applicationList/useApplicationList';
+import { useApplicationListData } from '../components/applicationList/useApplicationListData';
 import { $BackgroundWrapper } from '../components/layout/Layout';
-
-export const allApplicationStatuses: APPLICATION_STATUSES[] = [
-  APPLICATION_STATUSES.RECEIVED,
-  APPLICATION_STATUSES.HANDLING,
-  APPLICATION_STATUSES.INFO_REQUIRED,
-  APPLICATION_STATUSES.ACCEPTED,
-  APPLICATION_STATUSES.DRAFT,
-  APPLICATION_STATUSES.REJECTED,
-];
+import { ALL_APPLICATION_STATUSES } from '../constants';
 
 const ApplicantIndex: NextPage = () => {
   const {
@@ -45,17 +39,27 @@ const ApplicantIndex: NextPage = () => {
 
   const translationBase = 'common:applications.list.headings';
 
-  const { t, list } = useApplicationList(allApplicationStatuses, true);
+  const { list, shouldShowSkeleton } = useApplicationListData(
+    ALL_APPLICATION_STATUSES,
+    true
+  );
+  const { t } = useApplicationList();
+
+  const getHeadingTranslation = (
+    headingStatus: APPLICATION_STATUSES | 'all'
+  ): string => t(`${translationBase}.${headingStatus}`);
+
+  const getTabCount = (statuses: APPLICATION_STATUSES[]): number =>
+    list.filter((app: ApplicationListItemData) => statuses.includes(app.status))
+      .length;
 
   const getListHeadingByStatus = (
     headingStatus: APPLICATION_STATUSES | 'all',
     statuses: APPLICATION_STATUSES[]
   ): string =>
     list && list?.length > 0
-      ? `${t(`${translationBase}.${headingStatus}`)} (${
-          list.filter((app) => statuses.includes(app.status)).length
-        })`
-      : `${t(`${translationBase}.${headingStatus}`)}`;
+      ? `${getHeadingTranslation(headingStatus)} (${getTabCount(statuses)})`
+      : getHeadingTranslation(headingStatus);
 
   return (
     <FrontPageProvider>
@@ -65,7 +69,7 @@ const ApplicantIndex: NextPage = () => {
           <Tabs theme={theme.components.tabs}>
             <Tabs.TabList style={{ marginBottom: 'var(--spacing-m)' }}>
               <Tabs.Tab>
-                {getListHeadingByStatus('all', allApplicationStatuses)}
+                {getListHeadingByStatus('all', ALL_APPLICATION_STATUSES)}
               </Tabs.Tab>
               <Tabs.Tab>
                 {getListHeadingByStatus(APPLICATION_STATUSES.DRAFT, [
@@ -97,13 +101,19 @@ const ApplicantIndex: NextPage = () => {
 
             <Tabs.TabPanel>
               <ApplicationList
+                isLoading={shouldShowSkeleton}
+                list={list}
                 heading={t(`${translationBase}.all`)}
-                status={allApplicationStatuses}
+                status={ALL_APPLICATION_STATUSES}
               />
             </Tabs.TabPanel>
 
             <Tabs.TabPanel>
               <ApplicationList
+                isLoading={shouldShowSkeleton}
+                list={list.filter((app) =>
+                  [APPLICATION_STATUSES.DRAFT].includes(app.status)
+                )}
                 heading={t(`${translationBase}.draft`)}
                 status={[APPLICATION_STATUSES.DRAFT]}
               />
@@ -111,6 +121,10 @@ const ApplicantIndex: NextPage = () => {
 
             <Tabs.TabPanel>
               <ApplicationList
+                isLoading={shouldShowSkeleton}
+                list={list.filter((app) =>
+                  [APPLICATION_STATUSES.RECEIVED].includes(app.status)
+                )}
                 heading={t(`${translationBase}.received`)}
                 status={[APPLICATION_STATUSES.RECEIVED]}
               />
@@ -118,10 +132,18 @@ const ApplicantIndex: NextPage = () => {
 
             <Tabs.TabPanel>
               <ApplicationList
+                isLoading={shouldShowSkeleton}
+                list={list.filter((app) =>
+                  [APPLICATION_STATUSES.HANDLING].includes(app.status)
+                )}
                 heading={t(`${translationBase}.handling`)}
                 status={[APPLICATION_STATUSES.HANDLING]}
               />
               <ApplicationList
+                isLoading={shouldShowSkeleton}
+                list={list.filter((app) =>
+                  [APPLICATION_STATUSES.INFO_REQUIRED].includes(app.status)
+                )}
                 heading={t(`${translationBase}.infoRequired`)}
                 status={[APPLICATION_STATUSES.INFO_REQUIRED]}
               />
